feat: add query parameter to URL
